+++ inst/NEWS.md
@@ -7,6 +7,10 @@
 ## ENHANCEMENTS
 - Better sorting of oncoplot with `collapsePathway`
 - Changed default background for oncoplot from `gray` to `#ecf0f1`
+- Changed default signature database to SBS_v3.4 (from legacy)
+
+## NEW FUNCTIONS
+- `gisticCompare()` for comparing two GISTIC objects
 
 # CHANGES IN VERSION 2.18.0
 (Bioconductor release branch)